+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This file contains the list of changes made to the Joulescope driver.
 * Improved JS110 driver to provide current_range, gpi0, gpi1.
 * Fixed power, charge and energy computation.
   Power fixes require FPGA 0.2.6 upgrade.
+* Fixed JS110 stop/start streaming.
 
 
 ## 0.2.6